# src/swagger.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: to1step api docs
  description: to1step backend api 문서
  contact:
    email: gkqkehs0321@gmail.com
...
paths:
...
  /v1/stores/:
    post:
      tags:
        - StoreController
      summary: 가게 생성
      description: 새로운 가게를 생성하기
      parameters:
        - name: Authorization
          in: header
      requestBody:
        description: CreateStoreDto
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStoreDto'
      responses:
        '200':
          description: 가게 생성 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
                    description: 가게 생성 성공
                    example: true
  /v1/stores/me:
    get:
      tags:
        - StoreController
      summary: 내 가게 가져오기
      description: 내가 생성한 가게들 가져오기
      parameters:
        - name: Authorization
          in: header
      responses:
        '200':
          description: 가게 정보들
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetStoreDto'
  /v1/stores/location:
    get:
      tags:
        - StoreController
      summary: 내 주위 가게들 가져오기
      description: 내 주위 가게들 가져오기
      parameters:
        - name: region
          in: query
          description: 해당 지역 두 단어 ex) 서울시 마포구
      responses:
        '200':
          description: 가게 정보들
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetStoreDto'
  '/v1/stores/{storeUUID}':
    get:
      tags:
        - StoreController
      summary: 특정 가게 정보 가져오기
      description: 가게 정보 가져오기
      parameters:
        - name: storeUUID
          in: path
          description: 가게 UUID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 가게 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntireStoreDto'
    patch:
      tags:
        - StoreController
      summary: 가게 정보 수정
      description: 가게 정보 수정하기
      parameters:
        - name: Authorization
          in: header
          required: true
        - name: storeUUID
          in: path
          description: 가게 UUID
          required: true
          schema:
            type: string
      requestBody:
        description: UpdateStoreDto
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStoreDto'
      responses:
        '200':
          description: 가게 정보 수정 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
                    description: 가게 정보 수정 성공
                    example: true
    delete:
      tags:
        - StoreController
      summary: 가게 삭제하기
      description: 가게를 삭제하기
      parameters:
        - name: Authorization
          in: header
          required: true
        - name: storeUUID
          in: path
          description: 가게 UUID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 가게 삭제 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
                    description: 가게 삭제 성공
                    example: true
  '/v1/stores/{storeUUID}/like':
    post:
      tags:
        - StoreController
      summary: 가게 좋아요
      description: 특정 가게 좋아요 하기
      parameters:
        - name: Authorization
          in: header
          required: true
        - name: storeUUID
          in: path
          description: 가게 UUID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 가게 좋아요 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
                    description: 가게 좋아요 성공
                    example: true
    delete:
      tags:
        - StoreController
      summary: 가게 좋아요 취소
      description: 특정 가게 좋아요 취소 하기
      parameters:
        - name: Authorization
          in: header
          required: true
        - name: storeUUID
          in: path
          description: 가게 UUID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 가게 좋아요 취소 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
                    description: 가게 좋아요 취소 성공
                    example: true
  '/v1/stores/{storeUUID}/review':
    post:
      tags:
        - StoreController
      summary: 가게 리뷰 작성
      description: 가게에 리뷰 작성
      parameters:
        - name: Authorization
          in: header
          required: true
        - name: storeUUID
          in: path
          description: 가게 UUID
          required: true
          schema:
            type: string
      requestBody:
        description: CreateStoreReviewDto
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStoreReviewDto'
      responses:
        '200':
          description: 가게 리뷰 작성 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
                    description: 가게 리뷰 작성 성공
                    example: true
  '/v1/stores/{storeUUID}/reviews/{storeReviewUUID}':
    patch:
      tags:
        - StoreController
      summary: 가게 리뷰 수정
      description: 가게에 리뷰 수정
      parameters:
        - name: Authorization
          in: header
          required: true
        - name: storeUUID
          in: path
          description: 가게 UUID
          required: true
          schema:
            type: string
        - name: storeReviewUUID
          in: path
          description: 가게리뷰 UUID
          required: true
          schema:
            type: string
      requestBody:
        description: UpdateStoreReviewDto
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStoreReviewDto'
      responses:
        '200':
          description: 가게 리뷰 수정 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
                    description: 가게 리뷰 수정 성공
                    example: true
    delete:
      tags:
        - StoreController
      summary: 가게 리뷰 삭제
      description: 가게에 리뷰 삭제
      parameters:
        - name: Authorization
          in: header
          required: true
        - name: storeUUID
          in: path
          description: 가게 UUID
          required: true
          schema:
            type: string
        - name: storeReviewUUID
          in: path
          description: 가게리뷰 UUID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 가게 리뷰 삭제 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
                    description: 가게 리뷰 삭제 성공
                    example: true
  '/v1/stores/{storeUUID}/image':
    post:
      tags:
        - StoreController
      summary: 가게 리뷰 이미지 생성
      description: 가게에 리뷰 이미지 생성
      parameters:
        - name: Authorization
          in: header
          required: true
        - name: storeUUID
          in: path
          description: 가게 UUID
          required: true
          schema:
            type: string
      requestBody:
        description: CreateCourseReviewImageDto
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseReviewImageDto'
      responses:
        '200':
          description: 가게 이미지 생성 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: boolean
                    description: 가게 이미지 생성 성공
                    example: true
...
components:
  schemas:
...
    CreateStoreDto:
      type: object
      required:
        - name
        - category
        - description
        - location
        - coordinates
      properties:
        name:
          type: string
          description: 가게 이름
          example: 행보칸 카페
        category:
          type: number
          description: 가게 카테고리 0 카페 / 1 식당 / 2 공원
          example: 0
        description:
          type: string
          description: 가게에 대한 설명
          example: 커피가 맛있는 감성있는 카페
        location:
          type: string
          description: 가게 주소
          example: 경인남길 134-1
        coordinates:
          type: array
          description: 가게 좌표
          example:
            - 123.45
            - 12.34
        representImage:
          type: string
          description: 가게 대표 사진 경로
          example: storeImage.png
        tags:
          type: array
          description: 생성할 태그
          example:
            - 맛집
            - 인스타감성
            - 야경
        startTime:
          type: string
          description: 가게 오픈 시간
          example: 아침 9시
        endTime:
          type: string
          description: 가게 마감 시간
          example: 저녁 10시
    GetStoreDto:
      type: object
      properties:
        name:
          type: string
          description: 가게 이름
          example: 행보칸 카페
        category:
          type: number
          description: 가게 카테고리 0 카페 / 1 식당 / 2 공원
          example: 0
        description:
          type: string
          description: 가게에 대한 설명
          example: 커피가 맛있는 감성있는 카페
        location:
          type: string
          description: 가게 주소
          example: 경인남길 134-1
        coordinates:
          type: array
          description: 가게 좌표
          example:
            - 123.45
            - 12.34
        representImage:
          type: string
          description: 가게 대표 사진 경로
          example: storeImage.png
        tags:
          type: array
          description: 생성할 태그
          example:
            - 맛집
            - 인스타감성
            - 야경
        startTime:
          type: string
          description: 가게 오픈 시간
          example: 아침 9시
        endTime:
          type: string
          description: 가게 마감 시간
          example: 저녁 10시
    EntireStoreDto:
      type: object
      properties:
        uuid:
          type: string
          description: 가게 UUID
          example: 1210c3f2-4f68-4e85-94a7-81dc3d764393
        name:
          type: string
          description: 가게 이름
          example: 행보칸 카페
        category:
          type: number
          description: 가게 카테고리 0 카페 / 1 식당 / 2 공원
          example: 0
        description:
          type: string
          description: 가게에 대한 설명
          example: 커피가 맛있는 감성있는 카페
        location:
          type: string
          description: 가게 주소
          example: 경인남길 134-1
        coordinates:
          type: array
          description: 가게 좌표
          example:
            - 123.45
            - 12.34
        representImage:
          type: string
          description: 가게 대표 사진 경로
          example: storeImage.png
        tags:
          type: array
          description: 생성할 태그
          example:
            - 맛집
            - 인스타감성
            - 야경
        startTime:
          type: string
          description: 가게 오픈 시간
          example: 아침 9시
        endTime:
          type: string
          description: 가게 마감 시간
          example: 저녁 10시
        storeReviews:
          type: array
          description: 가게 리뷰들
          items:
            properties:
              uuid:
                type: string
                description: 해당 가게 UUID
                example: 1210c3f2-4f68-4e85-94a7-81dc3d764393
              user:
                type: string
                description: 작성한 유저 UUID
                example: 1210c3f2-4f68-4e85-94a7-81dc3d764393
              review:
                type: string
                description: 리뷰 내용
                example: 정말 맛있어요
        reviewCount:
          type: number
          description: 리뷰 개수
          example: 12
        likeCount:
          type: number
          description: 좋아요 개수
          example: 13
        iLike:
          type: boolean
          description: 내가 해당 가게에 좋아요 누른 여부
          example: true
    UpdateStoreDto:
      type: object
      required:
        - name
        - category
        - description
        - location
        - coordinates
      properties:
        name:
          type: string
          description: 가게 이름
          example: 행보칸 카페
        category:
          type: number
          description: 가게 카테고리 0 카페 / 1 식당 / 2 공원
          example: 0
        description:
          type: string
          description: 가게에 대한 설명
          example: 커피가 맛있는 감성있는 카페
        location:
          type: string
          description: 가게 주소
          example: 경인남길 134-1
        coordinates:
          type: array
          description: 가게 좌표
          example:
            - 123.45
            - 12.34
        representImage:
          type: string
          description: 가게 대표 사진 경로
          example: storeImage.png
        tags:
          type: array
          description: 생성할 태그
          example:
            - 맛집
            - 인스타감성
            - 야경
        startTime:
          type: string
          description: 가게 오픈 시간
          example: 아침 9시
        endTime:
          type: string
          description: 가게 마감 시간
          example: 저녁 10시
    CreateStoreReviewDto:
      type: object
      required:
        - review
      properties:
        review:
          type: string
          description: 리뷰 내용
          example: 너무 맛있어요
    UpdateStoreReviewDto:
      type: object
      required:
        - review
      properties:
        review:
          type: string
          description: 리뷰 내용
          example: 너무 맛없어요
    CreateCourseReviewImageDto:
      type: object
      required:
        - src
      properties:
        src:
          type: string
          description: 이미지 주소
...